+++ lib/karafka/pro/routing/features/virtual_partitions/contracts/topic.rb
@@ -26,6 +26,7 @@ module Karafka
                 required(:reducer) { |val| val.respond_to?(:call) }
                 required(:max_partitions) { |val| val.is_a?(Integer) && val >= 1 }
                 required(:offset_metadata_strategy) { |val| %i[exact current].include?(val) }
+                required(:distribution) { |val| %i[consistent balanced].include?(val) }
               end
 
               # When virtual partitions are defined, partitioner needs to respond to `#call` and it